[move/#16] enum class type 디렉토리로 이동
